Update __init__.py
Adding `pricing_plan` key to billing dict. When creating your Product in Stripe Billing, you'll assign it a Pricing Plan. That `price_` object in the Product details is the value for the `pricing_plan` key. This will probably be useful in the future if you wanted to extend this to cover multiple products, or multiple pricing plans related to the product. In the short term, it just gets you better alignment to Stripe's product model.